% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psbcSpeedUp-package.R
\docType{package}
\name{psbcSpeedUp-package}
\alias{psbcSpeedUp-package}
\title{psbcSpeedUp: Penalized Semiparametric Bayesian Cox Models}
\description{
Algorithms to speed up the Bayesian Lasso Cox model (Lee et al., Int J Biostat, 2011 \doi{10.2202/1557-4679.1301}) and the Bayesian Lasso Cox with mandatory variables (Zucknick et al. Biometrical J, 2015 \doi{10.1002/bimj.201400160}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ocbe-uio/psbcSpeedUp}
  \item Report bugs at \url{https://github.com/ocbe-uio/psbcSpeedUp/issues}
}

}
\author{
\strong{Maintainer}: Zhi Zhao \email{zhi.zhao@medisin.uio.no}

Authors:
\itemize{
  \item Manuela Zucknick
  \item Maral Saadati
  \item Axel Benner
}

}
\keyword{internal}
